+++ core/src/main/kotlin/br/com/guiabolso/events/json/LocalDateSerializer.kt
@@ -1,0 +1,23 @@
+package br.com.guiabolso.events.json
+
+import com.google.gson.JsonDeserializationContext
+import com.google.gson.JsonDeserializer
+import com.google.gson.JsonElement
+import com.google.gson.JsonPrimitive
+import com.google.gson.JsonSerializationContext
+import com.google.gson.JsonSerializer
+import java.lang.reflect.Type
+import java.time.LocalDateTime
+import java.time.format.DateTimeFormatter
+
+object LocalDateSerializer : JsonSerializer<LocalDateTime>, JsonDeserializer<LocalDateTime> {
+
+    override fun serialize(src: LocalDateTime, typeOfSrc: Type, context: JsonSerializationContext): JsonElement {
+        return JsonPrimitive(src.format(DateTimeFormatter.ISO_LOCAL_DATE))
+    }
+
+    override fun deserialize(json: JsonElement, typeOfT: Type, context: JsonDeserializationContext): LocalDateTime {
+        return LocalDateTime.parse(json.asString, DateTimeFormatter.ISO_LOCAL_DATE)
+    }
+
+}