+++ core/src/main/kotlin/br/com/guiabolso/events/json/MapperHolder.kt
@@ -1,11 +1,17 @@
 package br.com.guiabolso.events.json
 
 import com.google.gson.GsonBuilder
+import java.time.Instant
+import java.time.LocalDate
+import java.time.LocalDateTime
 
 object MapperHolder {
 
     @JvmField
     var mapper = GsonBuilder()
+        .registerTypeHierarchyAdapter(LocalDate::class.java, LocalDateSerializer)
+        .registerTypeHierarchyAdapter(LocalDateTime::class.java, LocalDateTimeSerializer)
+        .registerTypeHierarchyAdapter(Instant::class.java, InstantSerializer)
         .serializeNulls()
         .create()!!